increased the size of an internal variable as the Ai could theoretically have more then 256 lumberjacks

+++ src/ai/defaultai.cc
@@ -5620,7 +5620,7 @@ BuildingNecessity DefaultAI::check_building_necessity(BuildingObserver& bo,
 		} else if (bo.is(BuildingAttribute::kRanger)) {
 
 			// making sure we have one completed supported lumberjack
-			uint8_t supported_lumberjack_built = 0;
+			uint16_t supported_lumberjack_built = 0;
 			for (const auto& supported_building : bo.supported_producers) {
 				supported_lumberjack_built += get_building_observer(supported_building.first).cnt_built;
 			}
